+++ include/compTask.hpp
@@ -9,14 +9,14 @@
 struct ComputeTask {
   //
   struct Face {
-    Eigen::ArrayX3f points;
+    Eigen::ArrayX3d points;
     std::size_t faceIdx;
-    Eigen::Array3f centrePoint;
-    float area;
+    Eigen::Array3d centrePoint;
+    double area;
   };
   Face face;
   std::vector<std::size_t> indices;
-  Eigen::ArrayX3f points;
+  Eigen::ArrayX3d points;
 };
 
 using ComputeTaskPair =
@@ -24,7 +24,7 @@ using ComputeTaskPair =
 // Ctor
 template <SurfaceType Surface>
 ComputeTask createInfluenceComputeTask(const PanelGeometry<Surface> &panelGeo,
-                                       const EvalPoints<float> &evalPoints,
+                                       const EvalPoints<double> &evalPoints,
                                        std::size_t faceIdx) {
   ComputeTask compTask;
 
@@ -40,7 +40,7 @@ ComputeTask createInfluenceComputeTask(const PanelGeometry<Surface> &panelGeo,
 template <SurfaceType Surface>
 void createInfluenceComputeTask(ComputeTask &compTask,
                                 const PanelGeometry<Surface> &panelGeo,
-                                const EvalPoints<float> &evalPoints,
+                                const EvalPoints<double> &evalPoints,
                                 std::size_t faceIdx) {
 
   compTask.face.faceIdx = faceIdx;
@@ -51,8 +51,8 @@ void createInfluenceComputeTask(ComputeTask &compTask,
 }
 
 ComputeTaskPair makeComputeTasksPairImpl(const PanelGeometryPair &panelGeometry,
-                                         const EvalPoints<float> &evalPoints);
+                                         const EvalPoints<double> &evalPoints);
 
 std::vector<ComputeTaskPair>
 makeComputeTaskPairs(std::span<PanelGeometryPair> panelGeometries,
-                     const EvalPoints<float> &evalPoints);
+                     const EvalPoints<double> &evalPoints);
