+++ include/singularity/iconst_sing.hpp
@@ -3,13 +3,13 @@
 #include <Eigen/Core>
 
 template <typename Derived> struct IConstant3dSingularity {
-  static Eigen::ArrayXf calcInfluence(const ComputeTask &compTask) {
+  static Eigen::ArrayXd calcInfluence(const ComputeTask &compTask) {
     return Derived::calcInfluenceImpl(compTask);
   }
-  static Eigen::ArrayXf calcInfluenceFar(const ComputeTask &compTask) {
+  static Eigen::ArrayXd calcInfluenceFar(const ComputeTask &compTask) {
     return Derived::calcInfluenceFarImpl(compTask);
   }
-  static float calcSelfInfluence(const ComputeTask &compTask) {
+  static double calcSelfInfluence(const ComputeTask &compTask) {
     return Derived::calcSelfInfluenceImpl(compTask);
   }
 };
