+++ include/solver/isolver.hpp
@@ -3,6 +3,6 @@
 
 struct ISolver {
   virtual ~ISolver() = default;
-  virtual Eigen::VectorXf solve(const Eigen::MatrixXf &lhs,
-                                const Eigen::VectorXf &rhs, float tol=1e-6, std::size_t maxit = 10) = 0;
+  virtual Eigen::VectorXd solve(const Eigen::MatrixXd &lhs,
+                                const Eigen::VectorXd &rhs, double tol=1e-6, std::size_t maxit = 10) = 0;
 };
